requirements.txt
Pillow==10.1.0
PyQt6==6.6.0
PyQt6-sip==13.6.0
tqdm==4.66.1
